requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncssh==2.21.0
attrs==25.3.0
audioop-lts==0.2.1
Authlib==1.6.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
bs4==0.0.2
build==1.2.2.post1
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.3
cryptography==45.0.5
curl_cffi==0.12.0
cycler==0.12.1
cymem==2.0.11
discord-protos==0.0.2
discord.py @ git+https://github.com/Rapptz/discord.py@4496df79d047b7c26f0413c8a34086d309b96904
discord.py-self @ git+https://github.com/dolfies/discord.py-self@7f2e278605a5137e93b7c4a24cdd87e77cbba5c1
distro==1.9.0
edge-tts==7.2.0
et_xmlfile==2.0.0
fastapi==0.116.1
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
multidict==6.4.4
murmurhash==1.0.13
mutagen==1.47.0
numpy==2.3.1
openai==1.97.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
preshed==3.0.10
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyNaCl==1.5.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
#PyQt5==5.15.11

#PyQt5-Qt5==5.15.17
#PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.2
tabulate==0.9.0
thinc==8.3.6
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
yt-dlp==2025.8.27
